requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
azure-appconfiguration==1.7.1
azure-appconfiguration-provider==2.0.0
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.21.0
azure-keyvault-secrets==4.10.0
azure-storage-blob==12.25.0
azure-storage-queue==12.13.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.10
cryptography==46.0.5
dateparser==1.2.0
Deprecated==1.2.15
Django==5.1.15
django-cors-headers==4.6.0
django-filter==24.3
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
IMAPClient==3.0.1
importlib_metadata==8.5.0
inflection==0.5.1
iniconfig==2.0.0
invoice2data==0.4.5
isodate==0.7.2
itypes==1.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.2.0
msrest==0.7.1
mssql-django==1.6
multidict==6.1.0
numpy==2.3.1
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
opencensus-ext-django==0.8.0
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.8.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
pillow==12.1.1
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.1
proto-plus==1.25.0
psutil==6.1.1
psycopg==3.2.9
psycopg2==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.6.2
pyodbc==5.2.0
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.6.3
whitenoise==6.9.0
wrapt==1.17.2
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0